+++ CHANGELOG.md
@@ -8,3 +8,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added command `Start-PSResourceGetBootstrap`.
+- Added bootstrap script as a release asset.